+++ src/helpers/dom/element.js
@@ -861,6 +861,22 @@ export function getSelectionText() {
 }
 
 /**
+ * Cross-platform helper to clear text selection.
+ */
+export function clearTextSelection() {
+  // http://stackoverflow.com/questions/3169786/clear-text-selection-with-javascript
+  if (window.getSelection) {
+    if (window.getSelection().empty) { // Chrome
+      window.getSelection().empty();
+    } else if (window.getSelection().removeAllRanges) { // Firefox
+      window.getSelection().removeAllRanges();
+    }
+  } else if (document.selection) { // IE?
+    document.selection.empty();
+  }
+}
+
+/**
  * Sets caret position in text input.
  *
  * @author http://blog.vishalon.net/index.php/javascript-getting-and-setting-caret-position-in-textarea/
