+++ src/helpers/mixed.js
@@ -102,7 +102,7 @@ export function _injectProductInfo(key, element) {
       if (!warningMessage) {
         if (releaseTime > keyGenTime + 1) {
           warningMessage = toSingleLine`
-          Your license key of Handsontable Pro has expired.‌‌‌‌
+          Your license key of Handsontable Pro has expired.‌‌‌‌ 
           Renew your maintenance plan at https://handsontable.com or downgrade to the previous version of the software.
           `;
         }
@@ -152,19 +152,3 @@ function _checkKeySchema(v) {
   return p === z;
 }
 /* eslint-enable */
-
-/**
- * Cross-platform helper to clear text selection.
- */
-export function clearTextSelection() {
-  // http://stackoverflow.com/questions/3169786/clear-text-selection-with-javascript
-  if (window.getSelection) {
-    if (window.getSelection().empty) { // Chrome
-      window.getSelection().empty();
-    } else if (window.getSelection().removeAllRanges) { // Firefox
-      window.getSelection().removeAllRanges();
-    }
-  } else if (document.selection) { // IE?
-    document.selection.empty();
-  }
-}
